Test C program to print a division sign.
I'm not sure how portable this is, but it looks decent on my laptop.

In the future, I'd be interested in testing this on linux, and then
eventually on Windows (which I think will need some work).

To test on other platforms, I really need to run with the GUI directly
on those platforms, rather than say via ssh, where I'd just be using my
own terminal all over again. I want to make sure this works on
local-to-platform terminals.