Add solver.py, which can solve puzzles (!).
This doesn't solve the way a human would. But cpus are fast enough that
this finds solutions for the puzzle sizes we play with.

In the future, I hope to add solving capabilities that think "like a
human," and can thus explain a solution and grade the difficulty of a
puzzle.

Note that the current solve could easily be expanded to verify either of
these two cases:
* There are no solutions; or
* there are multiple solutions.