Able to save/load/display a solution.
In the future it seems like a nice idea to hide the solution by default,
and only display it when requested. For example, maybe you get a puzzle
file from a friend. As a file format, it's nice to enable the solution
to be included. As a human solving a puzzle, though, you don't want to
see spoilers by default!